refactored to use with containers in travis
