+++ packages/sakuli-legacy/src/context/sahi/__mocks__/get-browser-list.function.ts
@@ -1,0 +1,8 @@
+import {isPresent} from "@sakuli/commons";
+
+export function getTestBrowserList(): [("firefox" | "chrome"), boolean][] {
+    return [
+        ['firefox', !isPresent(process.env.FIREFOX_WD_URL)],
+        ['chrome', !isPresent(process.env.CHROME_WD_URL)],
+    ];
+}
