+++ packages/sakuli-legacy/src/context/sahi/accessor/accessor-util.class.spec.ts
@@ -3,7 +3,7 @@ import {createTestEnv, createTestExecutionContextMock, mockHtml, TestEnvironment
 import {AccessorUtil} from "./accessor-util.class";
 import {RelationsResolver} from "../relations";
 import {TestExecutionContext} from "@sakuli/core";
-import {getTestBrowserList} from "../action/__mocks__/get-browser-list.function";
+import {getTestBrowserList} from "../__mocks__/get-browser-list.function";
 
 
 jest.setTimeout(15_000);
