--- packages/sakuli-legacy/src/context/sahi/action/__mocks__/get-browser-list.function.ts
@@ -1,6 +1,0 @@
-export function getTestBrowserList(): [("firefox" | "chrome"), boolean][] {
-    return [
-        ['firefox', true],
-        ['chrome', true],
-    ];
-}
