+++ packages/sakuli-legacy/src/context/sahi/action/focus-actions.function.spec.ts
@@ -3,7 +3,7 @@ import {By, ThenableWebDriver} from "selenium-webdriver";
 import {AccessorUtil} from "../accessor";
 import {RelationsResolver} from "../relations";
 import {focusActionApi} from "./focus-actions.function";
-import {getTestBrowserList} from "./__mocks__/get-browser-list.function";
+import {getTestBrowserList} from "../__mocks__/get-browser-list.function";
 
 jest.setTimeout(15_000);
 describe('focus-api', () => {
