+++ packages/sakuli-legacy/src/context/sahi/action/keyboard-actions.function.spec.ts
@@ -3,7 +3,7 @@ import {By, ThenableWebDriver} from "selenium-webdriver";
 import {keyboardActionApi} from "./keyboard-actions.function";
 import {AccessorUtil} from "../accessor";
 import {RelationsResolver} from "../relations";
-import {getTestBrowserList} from "./__mocks__/get-browser-list.function";
+import {getTestBrowserList} from "../__mocks__/get-browser-list.function";
 
 jest.setTimeout(15_000);
 describe('KeyboardActions', () => {
