+++ packages/sakuli-legacy/src/context/sahi/action/mouse-actions-api.function.spec.ts
@@ -3,8 +3,8 @@ import {By, ThenableWebDriver} from "selenium-webdriver";
 import {mouseActionApi} from "./mouse-actions-api.function";
 import {AccessorUtil} from "../accessor";
 import {RelationsResolver} from "../relations";
-import {getTestBrowserList} from "./__mocks__/get-browser-list.function";
 import {SahiElementQuery} from "../sahi-element.interface";
+import {getTestBrowserList} from "../__mocks__/get-browser-list.function";
 
 jest.setTimeout(25_000);
 describe('mouse-actions', () => {
