+++ packages/sakuli-legacy/src/context/sahi/helper/get-sibling-index.function.spec.ts
@@ -1,7 +1,7 @@
 import {By} from "selenium-webdriver";
 import {getSiblingIndex} from "./get-sibling-index.function";
 import {createTestEnv, mockHtml, TestEnvironment} from "../__mocks__";
-import {getTestBrowserList} from "../action/__mocks__/get-browser-list.function";
+import {getTestBrowserList} from "../__mocks__/get-browser-list.function";
 
 jest.setTimeout(10000);
 describe('getSiblingIndex', () => {
