+++ packages/sakuli-legacy/src/context/sahi/helper/is-equal.function.spec.ts
@@ -2,7 +2,7 @@ import {By} from "selenium-webdriver";
 import {distanceBetween} from "./distance-between.function";
 import {isEqual} from "./is-equal.function";
 import {createTestEnv, mockHtml, TestEnvironment} from "../__mocks__";
-import {getTestBrowserList} from "../action/__mocks__/get-browser-list.function";
+import {getTestBrowserList} from "../__mocks__/get-browser-list.function";
 
 jest.setTimeout(50000);
